requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
PyJWT==2.9.0
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.11.0
SQLAlchemy==2.0.36
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.2.1
